Added dimensionalitTrait (typeclass), and replaced some traits with abstract classes.
The dimensinonalityTrait allows us to look up the dimensionality directly
from the type, rather than having it to define in every class.
This required to replace the traits with an abstract class.
However, this anyway makes more sense for all classes that represent
ND-Entities, as they will never be mixed in anywhere.