Added visualization utilities using VTK and made cosmetic changes to enhance user experience.
First version of simple vtk tools to look at meshes and images.
Part of it is also a conversion routing for converting 2D and 3D images
to VTK, which can be useful in differnet contexts.

Besides these changes, we also made a couple of cosmetic changes, to make
the toolkit more easy to use from the command line.