Implemented new Sampling interface.
The trait Sampler is made more generic, such that it is not restricted
to boxed domains anymore. This change required quite a few changes in various
parts of the code. In addition to these changes, we have updated the tests and
made some of the tests pass again.
However, the tests are still not clean and need to be revised.