Added requirement for UniformSampler2D/3D that numberOfPoints need to be power of 2/3.
If the numberOfPoints of the uniformSampler is not a power of 2 (for 2d images),
or a power of 3 (for 3d images), the sampler will sample fewer points than requested,
which leads at various places in the code to wrong results. At some point we need to
handle that. Currently this requirement is a simple safeguard.