Refactoring of PixelValues into generic ScalarValues. Added ScalarMeshData.
In order to be able to use the same mechanism for handling scalars on
meshes, as for handling scalars in discrete images, we generalized the
PixelValue concept and called it now ScalarValue.
In addition, we also introduced a new class ScalarMeshData, that allow us
to augment a mesh with scalar data.