Made DiscreteDomain distinct type from Domain. Changed extent to corner in boxed domain.
Having DiscreteDomain as a subclass of domain can lead to the confusion that a discrete domain
is used where a continuous domain is required. As a discrete domain is undefined everywhere except on a
finite number of points, using a discrete domain where a continuous domain is expected does never make sense.

This commit also changes the name of BoxedDomain to BoxDomain and the corner (which was mistakenly referred to as extent)
is now correctly named.