Improved accuracy when computing nystrom approximation. Fixes #176.
There are now two variants of the Nystrom method. In the first variant the user does specify
the number of leading eigenfunctions that should be computed. In this case a RandomSVD is used for
computing the leading eigenvectors of the Kernel matrix. If the user does not specify the number of
eigenfunctions, the method chooses the number of eigenfunctions such that a certain relative error tolerance
is fulfilled.