Improved accuracy when computing nystrom approximation. Fixes #176.
The eigenvectors needed in the nystrom methods are always comuted to a certain accuracy, using the
pivoted cholesky decomposition. It is no longer possible to specify the number of basis functions.
As the basis function computation is lazy, it does not cost much to return to many, as they can later
be ignored by the calling method.