More general interface for Metric (#200)
The method takeDerivativeWRTToTransform returned a function
Point[_2D]=>Option[DenseVector[Double]], which can be used to compute
the derivative of the metric at every point.  This exposes too much
detail of the internal working of the metric to the outside (e.g. that
the final gradient value is obtained by integrating over all the
points of the domain). This is not the case for all metrics (i.e. the
MutualInformation metric cannot be computed by integrating over
points).

We therefore changed the interface and gradient computation mechanism,
such that the metric now computes the full gradient (including the
derivative of the transformation). The consequence is, that the metric
now needs to know about the TransformationSpace, in order to compute
this full gradient.