Made DiscreteDomain a type parameter in DiscreteField
A DiscreteField is defined using a DiscreteDomain and a set of values,
which correspond to the values at the points of the domain. So far the
information about the concrete type of the DiscreteDomain was not available
in the DiscreteField. It was only know that it was a subtype of DiscreteField.
By introducing the concrete Domain as a type parameter into DiscreteField,
we can retain this information. This makes it possible that methods defined
on discrete fields can exploit more information about the domain,
such as for example that it is defined on a regular grid.
A typical example is an interpolation method, which can be much
more efficiently implemented if the underlying domain is an image domain.

cleanup