corrected conversion of vtk scalar types.
The previous implementation of the conversion from vtk Scalar Arrays to the corresponding
Scalismo ScalarArray was based on casts of the array (using asInstanceOf). While this compiles,
the containing types is not actually converted and accessing it leads to classCastExceptions at
runtime. This commit does an explicit conversion of the elements, which fixes this problem.
The corresponding test was also improved, in order to capture this mistake.