removed explicit type bound Dim
Many classes in Scalismo are parametrized over the dimensionality. The
corresponding type parameter is bounded by Dim, to make sure that
nothing else than _1D, _2D, or _3D could be passed. As this is only
documentation, but makes many type signatures significantly longer and
more unreadable, it has been removed.