fixed bug in tetrahedralmeshConversion and introduced template option
The conversion method for converting a tetrahedral mesh to a
vtkUnstructuredGrid takes now an additional (optional) parameter
template. This parameter  makes it possible to specify an existing
tetrahedral mesh as a blueprint. The basic structure of the tetrahedral
mesh is then taken from this template and only the points are changed.
This greatly speeds up the conversion.

At the same time the conversion to vtk was simplified and a bug (wrong
allocation of cell types) was fixed.