Transformation refactoring (#350)
Huge refactoring of the transformations to make them much more convenient to use. In particular, it is now possible to create transformations directly instead of going via the transformation spaces. Transformation spaces are still there as a concept, but are mainly used as factories to generate transformations using a unified interface. 

The concept of RigidTransformation and SimilarityTransformation also changed. Before a RigidTransformation and SimilarityTransformation were concrete types of transformation. Now these are just traits, which are implemented by  a variety of different classes (e.g. Rotation, Translation, RotationThenTranslation, ...).

Besides the conceptual changes, this commit also changes the organisation of the transformation. They are now moved out of the registration and put into their own package.