Introduced MALA proposal
This commit introduces a proposal, which, when used in a Metropolis-Hastings algorithm, corresponds to the Metropolis-adjusted Langevin algorithm. As this is based on gradients, the mechanisms for handling gradients in the proposals had to be extended. In particular, the product evaluator had to be extended to properly combine evaluators with gradients.