Introduced SampleLense and added GaussianRandomWalkProposal (#360)
Introduced SampleLens and PartialProposalAdapter

This commit introduced the trait SampleLens, which is a general way to
access parts of a generic Sample for the mcmc framework. This mechanism
makes it possible to keep the samples completely generic, but still be able
to provide useful proposal generators. As an example, the GaussianRandomWalkProposal is introduced.
The commit also introduces the PartialProposalAdapter, which makes it possible to combine parts of 
proposals to larger ones. Again, the Samplelens is used here to focus on the parts.