replaced TypeTag with enumeration in Scalar class (#366)
The use of a TypeTag to distinguish Scalartypes was always an ugly hack,
which required the use of reflection. In this commit we completly
replace the TypeTag with an explicit enumeration of the possible
Scalartypes. The type is then passed along with the Scalar typeclass.

The original reason to make this change is the Scala 3 migration, where
TypeTags seem to cause problem. It is, however, also more elegant and
simplifies the type signature.