use direction vectors instead of transformation in StructuredPoints
Some image formats allow that images are not aligned to the standard
coordinate axis, but rotated by a given transformation. So far this was
handled by specifying an anisotropic transformation in the
structuredpointsdomain. In this commit we change this to use three
direction vectors instead. Specifying the direction vectors seems more
intuitive than specifying a transformation and it is also
what tools like Slicer and itkSnap report. Furthermore, by specifying
the direction vectors we can deal with reflections, without having to
introduce new transformations into the already complicated
transformation hierarchy.