corrected bug in nifti image when reading images with transform
The nifti format allows for the specification of a linear
transformation for transforming the intensities. If such
a transformation is specified, it makes most sense to
change the datatype to float, as otherwise the range of
the transformation might not be captured. This commit changes
the code such that whenever a transform is specified,
the datatype is fixed to float.