Improved image handling (#369)
This commit fixes several issues regarding scalismo handling of images. 

1. use direction vectors instead of transformation in StructuredPoints
2. support for images that are not aligned with the standard coordinate axis
3. reslicing with choosable interpolater when converting to vtk structured points
4. change of datatype of nifti images when a intensity transformation is specified
5. simplified IO interface

The first two items make it possible to read and represent images that are not aligned with the standard coordinate axis. Previously, these images always had to be resample on reading. With this change we can represent them directly. 

The third change fixes the problem that occurd when converging an labelmap to vtk. As the conversion routines used a cubic BSpline interpolation, the converted labelmap showed artifacts. This is fixed by allowing the user to specify the interpolator to be used when converting to vtk. 

The fourth change addresses a problem, which led to capped intensity ranges when a nifti transform was specifies for unsigned data types.  If an intensity transformation is specified  it makes most sense to
change the datatype to float, as otherwise the range of the transformation might not be captured. This commit changes
the code such that whenever a transform is specified, the datatype is fixed to float.

Finally the interface is simplified, in a way that the more advanced flags, like resample oblique and favourQForm are exposed only in the readNifti method and not in the more general readScalarImage method.