Added new MHProposal and proposal combinators (#375)
This commit adds a few classes that provide an abstraction layer on top of the sampling framework, which makes it more convenient to define proposals using the Metropolis-Hastings framework. 

The improvements include:
- An  MHSample, which provides a tag along with the parameters
- A new logger, which stores all accepted and rejected samples and provides methods for computing acceptance ratios
- An MHDistributionEvaluator, which provides the .cached method to cache proposal
- A GaussianRandomWalkProposal 

The main novelty is, however, a product proposal, which combines independent proposals, which are defined on part of the parameter space, into a larger proposal.  Together with the MHMixtureProposal, this makes it possible to specifying arbitrarily complex proposals from simpler ones.