Added optional scale argument to gaussian kernel
To create a meaningful deformation model using a GP with a Gaussian Kernel, two parameters need to be specified: the standard deviation, which models the smoothness, and the scale, which determines the amplitude (size) of the modelled functions. So far only the smoothness could be specified when constructing the Gaussian kernel. The scale was specified by multiplying the resulting kernel with a factor s^2. This commit adds an optional parameter to the Constructor, which makes it possible to add the factor already during construction.