use jhdf instead of the hdf5 native library for reading hdf5 files (#387)
There exist a number of scurity vulnerabilities in the older hdf5 implementation
that scalismo is using (see https://github.com/unibas-gravis/scalismo-native/issues/6). 
This commit changes all reading routines for hdf5 to use jhdf, which is a pure java 
implementation. Writing is still done using hdf5. The implementation is done such that 
much of the original structure is preserved. Once we get completely rid of the native hdf5 libraries, 
a refactoring should be done.