Statistical Model IO to hdf5-json (#395)
This commit introduces a major change of how statistical models
are written in Scalismo. Writing is now only supported in hdf.json
format. For this purpose, a separate scalismo-hdf5-json library is added as a
dependency. The commit also moves all the statistical model io files
into an own directory. Finally, instead of using strings to represent
path in HDF files, a new class HDFPath is used, which adds typesafety
to the method signatures.